feat: login as company

+++ internal/core/application/shared/dto/customer_account_dto.go
@@ -8,14 +8,14 @@ type CustomerAccountOutputDTO struct {
 	ID int
 	Name string
 	PermissionLevel int
-	TicketsLeft int
+	TicketsLeft int64
 }
 
 func NewCustomerAccountOutputDTO(
 	id int,
 	name string,
 	permissionLevel int,
-	ticketsLeft int,
+	ticketsLeft int64,
 ) *CustomerAccountOutputDTO {
 	return &CustomerAccountOutputDTO{
 		ID: id,
